+++ flutter_inappwebview_windows/windows/flutter_inappwebview_windows_plugin.cpp
@@ -6,7 +6,9 @@
 #include "headless_in_app_webview/headless_in_app_webview_manager.h"
 #include "in_app_browser/in_app_browser_manager.h"
 #include "in_app_webview/in_app_webview_manager.h"
+#include "platform_util.h"
 #include "webview_environment/webview_environment_manager.h"
+
 
 #pragma comment(lib, "Shlwapi.lib")
 #pragma comment(lib, "dxgi.lib")
@@ -32,14 +34,41 @@ namespace flutter_inappwebview_plugin
     inAppBrowserManager = std::make_unique<InAppBrowserManager>(this);
     headlessInAppWebViewManager = std::make_unique<HeadlessInAppWebViewManager>(this);
     cookieManager = std::make_unique<CookieManager>(this);
+    platformUtil = std::make_unique<PlatformUtil>(this);
+
+    window_proc_id = registrar->RegisterTopLevelWindowProcDelegate(
+      [this](HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
+      {
+        return HandleWindowProc(hWnd, message, wParam, lParam);
+      });
   }
 
   FlutterInappwebviewWindowsPlugin::~FlutterInappwebviewWindowsPlugin()
   {
+    if (registrar) {
+      registrar->UnregisterTopLevelWindowProcDelegate(window_proc_id);
+    }
     webViewEnvironmentManager = nullptr;
     inAppWebViewManager = nullptr;
     inAppBrowserManager = nullptr;
     headlessInAppWebViewManager = nullptr;
     cookieManager = nullptr;
+    platformUtil = nullptr;
+  }
+
+
+  std::optional<LRESULT> FlutterInappwebviewWindowsPlugin::HandleWindowProc(
+    HWND hWnd,
+    UINT message,
+    WPARAM wParam,
+    LPARAM lParam)
+  {
+    std::optional<LRESULT> result = std::nullopt;
+
+    if (platformUtil) {
+      result = platformUtil->HandleWindowProc(hWnd, message, wParam, lParam);
+    }
+
+    return result;
   }
 }