+++ flutter_inappwebview_windows/windows/flutter_inappwebview_windows_plugin.h
@@ -1,7 +1,6 @@
 #ifndef FLUTTER_PLUGIN_FLUTTER_INAPPWEBVIEW_PLUGIN_PLUGIN_H_
 #define FLUTTER_PLUGIN_FLUTTER_INAPPWEBVIEW_PLUGIN_PLUGIN_H_
 
-#include <flutter/method_channel.h>
 #include <flutter/plugin_registrar_windows.h>
 
 namespace flutter_inappwebview_plugin
@@ -11,6 +10,7 @@ namespace flutter_inappwebview_plugin
   class InAppBrowserManager;
   class HeadlessInAppWebViewManager;
   class CookieManager;
+  class PlatformUtil;
 
   class FlutterInappwebviewWindowsPlugin : public flutter::Plugin {
   public:
@@ -20,6 +20,7 @@ namespace flutter_inappwebview_plugin
     std::unique_ptr<InAppBrowserManager> inAppBrowserManager;
     std::unique_ptr<HeadlessInAppWebViewManager> headlessInAppWebViewManager;
     std::unique_ptr<CookieManager> cookieManager;
+    std::unique_ptr<PlatformUtil> platformUtil;
 
     static void RegisterWithRegistrar(flutter::PluginRegistrarWindows* registrar);
 
@@ -30,6 +31,14 @@ namespace flutter_inappwebview_plugin
     // Disallow copy and assign.
     FlutterInappwebviewWindowsPlugin(const FlutterInappwebviewWindowsPlugin&) = delete;
     FlutterInappwebviewWindowsPlugin& operator=(const FlutterInappwebviewWindowsPlugin&) = delete;
+  private:
+    // The ID of the WindowProc delegate registration.
+    int window_proc_id = -1;
+    std::optional<LRESULT> FlutterInappwebviewWindowsPlugin::HandleWindowProc(
+      HWND hWnd,
+      UINT message,
+      WPARAM wParam,
+      LPARAM lParam);
   };
 }
 #endif  // FLUTTER_PLUGIN_FLUTTER_INAPPWEBVIEW_PLUGIN_PLUGIN_H_