+++ flutter_inappwebview_windows/windows/in_app_webview/webview_channel_delegate.cpp
@@ -54,7 +54,7 @@ namespace flutter_inappwebview_plugin
   {
     decodeResult = [](const flutter::EncodableValue* value)
       {
-        return std::make_shared<PermissionResponse>(std::get<flutter::EncodableMap>(*value));
+        return value == nullptr || value->IsNull() ? std::optional<std::shared_ptr<PermissionResponse>>{} : std::make_shared<PermissionResponse>(std::get<flutter::EncodableMap>(*value));
       };
   }